+++ qord/models/users.py
@@ -27,7 +27,7 @@ from qord.models.base import BaseModel
 from qord.enums import DefaultAvatar
 from qord.internal.helpers import create_cdn_url, get_image_data, BASIC_EXTS
 from qord.internal.undefined import UNDEFINED
-from qord.internal.mixins import Comparable
+from qord.internal.mixins import Comparable, CreationTime
 
 import typing
 
@@ -44,7 +44,7 @@ __all__ = (
 )
 
 
-class User(BaseModel, Comparable):
+class User(BaseModel, Comparable, CreationTime):
     """Representation of a Discord user entity.
 
     Attributes